requirements.txt
aiofiles
dnspython
motor
python-dotenv
pyrofork
tgcrypto
uvloop
